Move `EstimationMethod` to upper layer
